requirements.txt

matplotlib>=3.0.0
numpy>=2.0.0
opencv-python>=4.0.0.0
pandas>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.13
PyYAML>=6.0.0
scipy>=1.0.0
seaborn>=0.11.0